fix: Verify and fix misconceptions for students 31-40 (200130-200139)
Fixed mismatches:
- Ramirez_Zoey_200134 Q4: Added DT003 (nextInt for doubles)
- Brown_John_200138 Q4: Added INPUT001 (commented out import)
- Jackson_Emma_200139 Q1: Added INPUT003 (not reading time input)

+++ authentic_seeded/Jackson_Emma_200139_Mixed/Q2.java
@@ -3,17 +3,16 @@
 // Assignment 2
 
 import java.util.Scanner;
-public class Q2
-{
-	public static void main(String[] args)
-	{
+
+public class Q2 {
+	public static void main(String[] args) {
 		Scanner scan = new Scanner(System.in);
 		System.out.print("Enter the driving distance in miles: ");
-		double distance = scan.nextDouble();
+		double distance = scan.nextInt();
 		System.out.print("Enter miles per gallon: ");
-		double milesPerGallon = scan.nextDouble();
+		double milesPerGallon = scan.nextInt();
 		System.out.print("Enter price in $ per gallon: ");
-		double price = scan.nextDouble();
+		double price = scan.nextInt();
 		double cost = (distance / milesPerGallon) * price;
 		System.out.println("The cost of driving is $" + cost);
 	}
